move duplicated code into MTE qstorage class
